[test] 플레이 리스트 조회 테스트 작성 및 빠진 로직 추가

+++ Project/src/test/java/com/iucyh/jjapcloud/repository/user/UserRepositoryTest.java
@@ -28,7 +28,7 @@ class UserRepositoryTest {
         user.setPassword("1234");
 
         User savedUser = repository.save(user);
-        int id = savedUser.getId();
+        long id = savedUser.getId();
 
         Optional<User> foundUser = repository.findById(id);
         assertThat(foundUser.isPresent()).isTrue();
@@ -40,7 +40,7 @@ class UserRepositoryTest {
     @DisplayName("유저 조회 성공")
     void find() {
         String uniqueString = UUID.randomUUID().toString();
-        int id = saveTestUser(uniqueString);
+        long id = saveTestUser(uniqueString);
         Optional<User> foundUser = repository.findById(id);
 
         assertThat(foundUser.isPresent()).isTrue();
@@ -52,7 +52,7 @@ class UserRepositoryTest {
     @DisplayName("유저 이메일로 조회 성공")
     void findByEmail() {
         String uniqueString = UUID.randomUUID().toString();
-        int id = saveTestUser(uniqueString);
+        long id = saveTestUser(uniqueString);
         Optional<User> foundUser = repository.findByEmail(uniqueString + "@abc.com");
 
         assertThat(foundUser.isPresent()).isTrue();
@@ -64,7 +64,7 @@ class UserRepositoryTest {
     @DisplayName("유저 조회 실패")
     void findFail() {
         String uniqueString = UUID.randomUUID().toString();
-        int id = saveTestUser(uniqueString);
+        long id = saveTestUser(uniqueString);
         Optional<User> foundUser = repository.findById(id + 1);
 
         assertThat(foundUser.isPresent()).isFalse();
@@ -74,7 +74,7 @@ class UserRepositoryTest {
     @DisplayName("유저 이메일로 조회 실패")
     void findByEmailFail() {
         String uniqueString = UUID.randomUUID().toString();
-        int id = saveTestUser(uniqueString);
+        long id = saveTestUser(uniqueString);
 
         Optional<User> foundUser = repository.findByEmail("fail@fail.com");
 
@@ -85,7 +85,7 @@ class UserRepositoryTest {
     @DisplayName("유저 업데이트 성공")
     void update() {
         String uniqueString = UUID.randomUUID().toString();
-        int id = saveTestUser(uniqueString);
+        long id = saveTestUser(uniqueString);
 
         User foundUser = repository.findById(id).orElseThrow();
         foundUser.setNickname("update");
@@ -101,7 +101,7 @@ class UserRepositoryTest {
     @DisplayName("유저 삭제 성공")
     void delete() {
         String uniqueString = UUID.randomUUID().toString();
-        int id = saveTestUser(uniqueString);
+        long id = saveTestUser(uniqueString);
         Optional<User> foundUser = repository.findById(id);
         assertThat(foundUser.isPresent()).isTrue();
 
@@ -111,7 +111,7 @@ class UserRepositoryTest {
         assertThat(deletedUser.isPresent()).isFalse();
     }
 
-    private int saveTestUser(String uniqueString) {
+    private long saveTestUser(String uniqueString) {
         User user = new User();
         user.setNickname(uniqueString);
         user.setPassword("123456");
